services/ocrmypdf/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pydantic==2.10.3
PyMuPDF==1.24.13
Pillow==11.0.0
img2pdf==0.5.1
